Refactors array declarations using new nonterm.
Both DirectAbstractDeclarator_c and DirectDeclarator_c are effected.

New Nonterminal is called `PostFixModifier_c`, and it allows us
to extend the syntax of array declarations, allowing the addition
of a "checked" qualifier on arrays, for instance, in the following way.

``
int a checked[10];
``

Also moves the definitions of DirectDeclarator_c, AbstractDeclarator_c,
and DirectAbstractDeclarator_c up to just after the definition of
Declarator_c, for readability.